requirements.txt
pillow
SimpleITK
imagecodecs
numpy
torch
transformers
ultralytics